[ledger] Store remote commit id to commit id mapping
When we add a commit, we add the mapping from its remote id to its local
id in the database. We provide a way to query this mapping too. Commits
that are in the process of being added but are not added yet are added
into a supplementary in-memory map that is also used to answer queries.

Change-Id: Ic1670796e31585e757c07b210bc9aa2d5281226d