[fuchsia_pkg_testing] Wait for blob to be creatable after closing channel
Dropping a FileProxy synchronously closes the zircon channel, but it is not guaranteed
that blobfs will respond to the channel closing before it responds to a request on a
separate channel to open the same blob. This means a test case that:
1. opens writable + truncates on channel 0
2. drops channel 0
3. opens writable on channel 1
can fail with ACCESS_DENIED in step 3, unless we wait.

Bug: 35723
Change-Id: Id31b4026bf414d9958540b99edff9abe26058029