[modular][inspect] solving a flake in inspect_session_test
The annotate module inspect hierarchy did not update in ~10%
of the tests. Moved the fidl clone of the module_data to
set_on_module_updated to accurately capture the module_data as it synced
with the story_storage and called ResetInspect.

This CL also breaks up the inspect_session_test into a mod hierarchy and
a story hierarchy to better isolate future test failures.

Bug: 37256

Change-Id: Icfca26fb9acaf9484cf35d859611cf0787d4a844