[zircon][build][kazoo] Add source_name_part to host_tool_action template
In host_tool_action_foreach, all invocations were using the same .d file
named only by target_name (which works file for host_tool_action).

So on subsequent runs, ninja would determine that the wrong output was
named in the .d file and so mark (one of the other) files as dirty.

Bug: 37505
Bug: 38488
Bug: 36581
Change-Id: Ia5830a6d935d6f70726e2b37c69a6f8319f8414e