[fbl][unique_ptr] Remove fbl::unique_ptr
Put a stake in the heart of fbl::unique_ptr once and for all.
Apologies about the reformatting side effects mixed up in this CL.
Extensive scripts were the only rational way to go about converting
all of this code to explicitly use std::unique_ptr, and fx format-code
had a necessary role to play there as well.

Change-Id: I2593fc4ab6395e0c0ba5e5cf2fbf46fcb70b9796