[json_parser] Enable json_parser target and tests for host, tidy up
This library contains json parsing utilities that are useful on host
and/or very similar to other library utilities that are useful on host.
This fixes up the build rules and code to build on all supported host
targets (avoiding strlcpy) and reuses directory enumeration / file
testing logic from src/lib/files/ and also tidies up the build rules
and namespace.

This also removes code in the appmgr integration test that attempted to
read configuration data from the path "/system/data/sysmgr" which no
longer exists. Previously, JSONParser::ParseFromDirectory would return
no error if it was unable to open a directory. This commit changes the
behavior to report an error in this case.

Bug: 36759

Change-Id: Ic7af0ac23181a9bff678204c68ed96beb9784f5b