[fbl] Remove fbl::unique_free_ptr and it's tests.
Now that std/ktl unique_ptr is available, there is no good reason to
continue to maintain a flavor of fbl::unique_ptr which calls free
instead of delete.

Furthermore, fbl::unique_ptr is nothing but an alias for
std::unique_ptr with the default deleter at this point.  There is no
good reason to maintain unit tests for an alias like this, so go ahead
and remove them.

In the near future, I plan to go through our code with a giant search
and replace and eliminate the need for the fbl::unique_ptr alias
completely, at which point in time we can eliminate the last traces of
fbl::unique_ptr

Change-Id: Idceeae385044cbb235b04b7fd8a2e3d27bd9f19e