[camera] add controller stream to camera_demo
This change add camera controller as an additional streaming source to
the camera manual test / demo. It also adds an extremely simple scenic
image node, populated with the text of the friendly name of the stream
source.

This change also fixes a bug in the legacy ISP driver that manifested in
the new camera_demo_test.

Test: ran the camera_demo_test
Change-Id: I082b1c47da8dc22182d49578452d30c8610160fd