[fvm] Convert to zxtest, isolated devmgr
Although there are more opportunities for cleanup
in the FVM integration tests, this patch aims to accomplish
a few major things:

1) Integration of an isolated devmgr. This will resolve
pending FVM flake issues.
2) Conversion to zxtest. This meshes with the isolated
devmgr usage in the harness.
3) Reduction in size and complexity of tests. Prefer
using smaller devices when possible.

As a side-effect of this change, support for running the FVM
tests on a "real block device" has been removed, in favor
of easier-to-understand, more hermetic testing.

Bug: 38562

Change-Id: Ifa050accd422f518e247ab2591a61ff15ac454e4