[fidl][llcpp] Support multiple in-flight async transactions
Adds |AsyncBinding| that implements channel to dispatcher binding for LLCPP
based servers allowing multiple in-flight |AsyncTransaction| objects.
The transactions handled via this binding can be safely sent to worker
threads that can reply in their own context, the binding keeps receiving
and dispatching messages on the channel while workers have not replied
to their requests.  Handling of a PEER_CLOSED from either the remote end of the
channel or the worker threads is serialized.

Bug: 7729

Test: runtests -t fidl-async-test-test
Change-Id: Ia360a53efb850e1e300d4ed10a663edc6702b3bf