[ledger][benchmarks] Remove expected_sample_count from add_new_page(_precached).tspec
Bug=38639

For these two benchmarks we measure the time to create a new Db. Each
page request necessarily leads to a new Db creation. Because of the
precaching mechanism though, we will always try to create one additional
Db. For this extra Db, we may or may not have enough time at the end
of the benchmark to create it, meaning that the expected sample count is
in [|page_requests|, |page_requests| + 1].

This CL removes the expected_sample_count to avoid flakes due to timing.

Change-Id: I4b86653b3a40fd0828bc22f39b66726c14e77606