[activity-service] Add fuchsia.ui.activity.Control interface
The Control protocol is a transitional API which allows an explicit
activity state to be passed into the Activity Service.

Once this protocol has been called once, the Activity Service will
always reflect the state set through later calls to this protocol. Input
to the Tracker API will continue to be received, but the StateProvider
protocol will only reflect state transitions occuring due to the Control
protocol.

This is transitional to support use cases where an existing subsystem is
responsible for determining the activity state of the system. Eventually
the goal is to have the Activity Service be the primary source of truth
of activity state on Fuchsia, at which point this API can be removed.

Bug: 38334
Change-Id: Ibd829ab204d61df33b173c057f187e1c806faf81