[media][tests] Address possible use-after-free.
The AudioPlugDetectorTest is adding some fake devices to a PseudoDir to
verify the plug detector is able to see these device nodes. The tests
were not removing these device nodes properly, so it's possible that
the vfs thread would access the device node that was on the stack after
the test body has completed, but before that thread is stopped in the
tests TearDown method.

MULTIPLY: `[{
  "target": "//src/media/audio/audio_core:audio_core_unittests",
  "total_runs": 200
}]`

Bug: 37173
Test: fx run-test audio_core_unittests
Change-Id: I6a14a55ced93aad96a954f580be932a3d269ea2d