[kernel][arm64][boot-shim] Do not extract kernel out of ZBI for qemu
Add a new switch for the qemu specific boot shim to set that causes the
kernel to not be extracted out of the ZBI as it lay in memory placed in
QEMU. This causes it to waste a little more memory (size of the kernel)
but greatly decreases the boot time on ThunderX2 based qemu-kvm hosts,
since the ZBI is no longer memmoved down to overwrite the kernel. Since
the boot shim runs with cache disabled, the performance is extremely
slow, on the order of 10s of seconds of startup time.

Also clean up the debug printfs a bit and add some code to print the 4
character type id of a ZBI section. All of that is disabled by #ifdef so
it's not currently built.

Tested: boot on qemu arm with and without kvm, and on at least one hw
target to make sure the boot shim is not modified for them.

Change-Id: Ifea0cbe2cd7e7a72b13908a102a2f9c6bd7be799