[kernel][mutex] Optimize spinning.
Now that ticks and clock monotonic units are guaranteed to be based on the same reference, convert
the spin timeout for kernel mutexes from clock monotonic to ticks.  This allows us to perform unit
conversion once at the start, instead of every time we compute now() while checking in the spin
loop.

Also, add a set of unit tests to verify that we are waiting for at _least_ the specified spin
timeout, and manually verify that we are close to our targets on hardware.  Note; we are much more
consistent on real hardware then we are on KVM QEMU hardware.

Tests: New unit kernel test.
Bug: 35437
Change-Id: Ic656a65bf97037599b3b25664d69caa74ece9f78