[json_parser] Consolidate JSON utilities in json_parser library
This moves the rapidjson_utils and peridot/lib/rapidjson utility
functions into the central json_parser library and unifies their
namespace. This also updates calling code to depend on the new
location and tightens up some dependency rules. This also removes a few
functions from the peridot/lib/rapidjson that appear to have no callers
or tests.

This moves the most commonly used type 'JSONParser' to the json_parser
namespace to match the library name and the rest of the utilities, but
leaves an alias behind in the namespace '::json' to make it easier to
transition.  I'll gradually update users of 'json::JSONParser' to the
new name in smaller patches and remove the alias once all are gone.

Bug: 36759
Change-Id: I03455d7be1ee5f82d25ccd68f0a067ca201ced2c