[pseudo-fs-mt] Support service nodes
While it is possible to write a simple implementatio of a
`DirectoryEntry`, implementation of a node that supports both
`NODE_REFERENCE` interaction and switch to an separate FIDL protocol is
non-trivial and requres a considerable amount of work.

It makes sense for the pseudo-fs-mt library to provide a shared
implementation.

BUG: 35109

Change-Id: I329be6ef752a1e8430632b1700f75155f4f9f70c