[cmgr][ws] Towards a WorkScheduler Dispatcher
This change implements a mostly-empty dispatcher of `WorkItem`s. As an
initial step, the `Dispatcher` is a trait implemented by `Realm`,
which owns the `root_exposed_dir` necessary to connect to a `Worker`
interface. A test impl of `Dispatcher` over `AbsoluteMoniker` limits the
apparatus needed for "dispatcher-free" `WorkScheduler` unit tests.

In the future, the production implementation of `Dispatcher` will likely
be an owner of a lazily instantiated FIDL client end to the `Worker`
interface that can be reused on periodic `WorkItem`s.

Bug: 35214
Bug: 35215

Change-Id: I616068e7a77c8cc74d7675d48b65451e52c6c8c3