[brcmfmac] Adding support for WPA1 in the driver
This change includes some refactoring of the authentication code in the
driver that was done by Josh. In addition, the driver retrieves WPA1 IE
(vendor specific IE) from beacons and passes it up to SME. And when
association request is received, the WPA1 IE is processed. If RSNE IE is
present it gets processed ahead of WPA1 IE.

Bug: 29793

Test: Verified that existing RSNE support is not affected and
functioning as expected. Also confirmed that Vendor IE (containing WPA1
IE) is being passed up to SME.

Change-Id: I8896e7b8838fa4e4cbe4ac6c0ad1591ec0d73125