[netemul] Test guest functionality in isolation.
`netemul_sandbox_test` tests all of netemul's features
except for guest control, so it is handy to run
`fx run-test netemul_sandbox_test` while iterating on
changes.  guest_test currently does not run on QEMU so it
gets in the way of local development.  This change moves
guest_test to its own test package to eliminate this issue.

Bug: 38618
Change-Id: I622ec6633abaa0dcd88bfe18144cc261e671290d