[battery_manager] refactor to separate manager
Migrate battery manager specific logic into a
dedicated module that we can build off of.

Role of main now isolated to initializing and
spawning the async tasks handled by the actual
battery manager.

This change also adds a call to get the up-to-date
power info as part of the periodic update, as it
was found that the power state could be stale.

Bug: go/fxb/7053
Test: ensure no regressions through refactor
      run existing tests -
      fx run-test battery_manager_tests
      fx run-test battery_manager_gotests

Change-Id: I0573062378027d6f0e5d69ed988b3ebead9891b0