[pseudo-fs-mt] EntryContainer trait
In order to support `Link` operations, we need to be able to get a
reference to a directory entry from any kind of directory that can
provide links, even read-only ones.  `EntryContainer` is a
representation of this capability - something that can provide an
`Arc<dyn DirectoryEntry>` given the entry name.

Change-Id: I42bfe29b75be3e5cb3ee97ad8891d3d3579f4758