[brcmfmac] Add thread names and move away from pthreads
All threads created using pthreads did not have names and it's
cumbersome to create attribute and set a name using pthreads. Having
thread names makes it easier to debug and analyze performance of the
driver.

This change moved to using the fuchsia threads library, specifically
thrd_create_with_name. This was already used in other parts of the
driver so no additional dependencies needed. The threads library seems
to be using pthreads under the hood so functionally this should be
identical.

The threads include header as a flag to enable the necessary functions
so the netbuf header had to change to set that flag or the include order
would prevent the declaration of that function.

Test: wlan-smoke-test, wlan-stress-test, manual testing

Change-Id: Ibe088ee02f86c7b912bffdd0e264b8de2239fea9