[fs] Move Connection::HandleFsSpecificMessage to Vnode
Because HandleFsSpecificMessage is load-bearing and cannot be removed,
an alternative way to support the equivalent functionality is to move
the function into Vnode. This way, the creation and serving of fs
connections could be entirely managed by the FIDL layer in a future CL.
Users of the library would not be able to override the type of the
connection on behalf of vfs.

It is certainly somewhat non-intuitive that a method call to handle
FIDL messages gets migrated to the Vnode class. However, upon inspecting
all the existing uses of that API, it turns out that all of them
immediately delegate into a corresponding Vnode method anyways (e.g.
GetAllocatedRegions from blobfs::DirectoryConnection calls into the
eponymous method in blobfs::Directory). As a workaround to support LLCPP
migration, this small inconsistency should be reasonable.

Bug: 34613
Bug: 39156

TEST: No observable behavior change
Change-Id: I4a8f19d81a1020044111c0e9a3e9fa4292ae4da0