[debugger] Send 512 bit x64 registers.
Currently Zircon doesn't support AVX-512 but the system API does. This
wires up the AVX-512 registers (which are considered by the protocol to
be the canonical vector registers) to the syscall.

This leaves the question about whether AVX or AVX-512 are supported
entirely to the kernel.

There should be no functional change since all valid data was already
being copied.

Change-Id: I0fabbfd2cefbba6f1dde7f83e6d6692326359dac