[kernel][vm] Unify VMAR allocation strategies
The three different allocation strategies can all be described in terms
of the general ASLR strategy by varying the levels of entropy.
Specifically the general allocator degrades to linear when asked to use
0 bits of entropy, and a specialized compact allocator is redundant now
that entropy can be arbitrarily tuned.

For the compact allocator entropy, the previous compact allocator had
approx. 8 bits of entropy and so whilst memory layouts will be
different, this change should be a no-op with regards to entropy.

Bug: 34546
Bug: 33764

Change-Id: I338df0efa1d7b246c484c7ce9618dc47de4ca8f5