[cm] Add validation of UseRunnerDecl blocks, add "source" attribute.
Add the ability to express UseRunnerDecl in FIDL "use" stanzas, and add
basic validation to ensure that blocks are well formed.

We additionally adjust "UseRunnerDecl" from just having the name of
a runner to use to also expressing the source of this name. This allows
disambiguating when the framework and a realm both provide a runner of
the same name, and matches the style of the other UseXxxDecl blocks.

BUG: 4761

Change-Id: I59f417011e304d42cbdf47047584ba76183301ba