[cmc] Factor out CML reference parsing.
Instead of exposing from the "cml" module regular expressions that
describe valid tokens, just expose functions for parsing the tokens.
This avoids duplication of the parsing logic in multiple places in
"validate.rs" and "compile.rs".

This is in prepartion for adding support for runners to the CML
validator/compiler, which will require yet another copy of the parsing
logic.

BUG: 4761
Change-Id: I6c364f3aa9896619c35a05e53e6775218fa8aedc