[fidlc] Goldens now include the C coding tables.
This will make it much easier to code-review diffs to the coding tables
generator & output, when the coding tables are modified to support the
old & v1 wire formats.

Also added an example "union sandwich" FIDL file that has useful data
structures to validate the size/offset/padding changes between the old
and v1 wire formats. Thanks to Pascal for coming up with these!

Bug: 38749
Bug: 39173

Test: fx exec garnet/go/src/fidl/compiler/backend/typestest/regen.sh

Change-Id: Ic16fb2351933175fe907ac907fc4ea54dedff62e