[brcmfmac] Firmware files for Astro & Sherlock are different.
This is a reland of 054fec0b4560f2b71449cbee7f76593508a65e94

Bug: 38410

This change copies the relevant files for the platform into the
firmware directory. During testing found that the iovar mpc fails
in Sherlock. So made the change to ignore the error and continue.

Test: Tested and Astro and Sherlock and verified the driver loads
correctly and is able to scan and connect to different APs.

Change-Id: Ie47e1fad6ebb4bffe3feb170ff5ea68eef6939c3