[debuglog][syscalls] Update zircon/sytem/core/devmgr/shared/fdio.{cc|h} to match src/devices/coordinator/fdio.{cc|h}
https://fuchsia-review.googlesource.com/c/fuchsia/+/330873 modified the
src/devices/coordinator copy of this code to take a root_resource for
creating a debuglog. It's a bit hairy to update the calls in fshost, so
as an interim step (so that the fdio.{cc|h} can be remerged?) make the
two copies match, and just pass in an invalid resource handle from the
callsites.

Bug: 32044
Change-Id: I47931deb754155ce707348d0e2f2cad3a089a7bf