[brcmfmac] Support dynamic interface creation
This commit introduces the WlanInterface class, which holds the Zircon
device and brcmfmac driver state for a WLAN interface instance.  The
Device class now supports multiple simultaneous interfaces by holding
multiple WlanInterface instances.

Note that at this point, only two interfaces are supported;
a client and an AP interface.  This should be sufficient for our current
purposes. Note that only one of them can exist at anytime as of now.
Will handle the simultaneous operation as a separate task.

Verified that the client interface can be deleted and an AP interface
can be created and a client is able to join to it.

Also, added support for creating and deleting an interface in SIM FW.

Test: Added a test, dynamic_if_test that creates and deletes a client
interface and an AP interface.

Change-Id: Ife73ee2fc47a818f23ddfd9f2b563558be062682