[zircon][digest] Add NodeDigest
This CL factors out the code used to calculate per-node digests in the
Merkle tree. "Node" are fixed-length chunks of data, e.g. every 8 KB
chunk of a blob in blobfs.

This is the first step in refactoring Merkle trees to make the code
easier to read and maintain, and the API more flexible.

Bug: 35678
Change-Id: I8959fbd0f620463f5794430f3048b1d15fb8c048