[debugger] Breakpoint Registering tests on Process.
This verifies existing behaviour, there weren't any test for it.
This required a bit of plumbing since the DebuggedProcess was a
ProcessMemoryAccessor, making it unable to mock for tests.

Now the accessor is a injectable dependency that will default to a
process-specific one that keeps the previous behaviour.

Renamed the process handle from process to handle, which is much
clearer.

TEST=Added registering breakpoint test.
     fx run-test debug_agent_tests

Change-Id: I679d846a2d1bf1b1a9210f8e0f5f7eb4ad6220c4