[FDF] Fix the flake in the suspend tests
The suspend tests are flaking because the test tear down trying to
teardown the vfs is racing with all the connection teardown on the async loop
are teared down. The class has the vfs after the loop and they get
destroyed in the reverse order. So the vfs shuts down first before async
loop tearsdown all the connections. In this change we ensuer, that the
async loop shutsdown first so that vfs gets teared down after.

Test: Running the tests with multiply[IN PROGRESS]. Stacktrace of the flake suggests
this is the issue. Hopefully this is it.

MULTIPLY: `[
  {
    "target": "//src/devices/coordinator:devcoordinator-test",
    "total_runs": 10
  }
]`

Bug:37462

Change-Id: Ifd6bd4c9aab8c76a4a445ab59eebc434b6dd95fc