[blobfs][fuzzit] Return an error message on zero sized buffers.
Currently zero sized buffers passed to ZSTDDecompress  will cause ZSTD to infinite
loop based on findings from the fuzzer. This change just detects this
condition and exits with an error message. I added a test case to
prevent this regression occuring in the future.

We may want to consider upstreaming a change to the third party library
in the future.

Bug: 34440
Change-Id: I226362c4a108ad7df6234703a229c73663d5c29a