[fidlc] New type shape tests for unions
Sandwiching a static-union in a struct to check the offset of various
fields and check that when all variants have an alignement of 4 or less,
no padding within the static-union is added.

Comparing to the post-migration world where the replaced inline
flexible-union header is of size 24, with an alignment of 8.

Bug: 38749 #done
Test: YES!
Change-Id: I3e8c795f9f144b9706159da0854ac9995633c593