[codec] ThreadSafeDeleter for Sync() callback
The FIDL SyncCallback must be deleted on the fidl_thread(), even if we
shut down the CodecImpl abrubtly and delete tasks queued to
StreamControl without running them.  This means the currying of
SyncCallback back to fidl_thread() doesn't happen, but thanks to
ThreadSafeDeleter, the deletion of SyncCallback itself (the logical one
that's been moved into) will still occur on fidl_thread().

BUG: 37790
Change-Id: Ic05bb0ee8cdba7af172a620a080adee5ec7f4d68