[debugger][exception_broker] Inform processes on limbo on zxdb.
The status command queries the processes waiting on limbo and shows it
on startup. This has to be hooked to the status command too.

Added the fuchsia.exception.ProcessLimbo service to the list of
processes the debug agent can connect to.

Fixed a bug on the exception broker where the handles could not be
transfered as they didn't have the ZX_RIGHT_TRANSFER right.

TEST=Activated limbo (Compile ExceptionBroker with use_limbo_ = true)
     1. fx shell -- crasher
     2. fx debug

     See output

Change-Id: I8de1870503218f6e833c0f1cc9d706274c612bc3