[fidl_codec] Add Encoder class
This pretty much does what it says on the tin. The class is usable via a
static method, but we keep the constructors private so we can manage the
lifecycle of the encoder (it's not to be used more than once).

There's no way to actually build value nodes yet, so this doesn't have
much use, but we can test it by re-encoding decoded messages in the
parser test.

Change-Id: I58b11c03648b273608a03a4a2f53a1c8169563e6