[debugger] Remove zxdb::Register, simplicification.
We had two objects named "Register" which was confusing and the
higher-level one wasn't adding any benefit other than a getter for an
int64 value. This getter is moved to the debug_ipc version, and the
higher-level client one is deleted.

The RegisterCategory struct was mostly adding a layer of indirection
without helping, so this is removed. RegisterCategory::Type is renamed
to RegisterCategory. The category stuff is now moved to the register
file.

The client RegisterSet is deleted and everything just uses vectors of
registers.

Register formatting and hex output is simplified to remove most error
handling. Since these functions are generating output for the user, they
can just write what they want in the format they want.

Change-Id: I6197ac3569ed3761240a56b4403c550c5e569492