[benchmarks] Fix storage performance tests by moving the "--quiet" argument
Move "--quiet" so that it's not passed to the storage performance
tests, since not all of them accept that argument.  (I didn't notice
that when I added the "--quiet" argument since the garnet-perf bot was
broken for a while due to problems with Zedboot in netboot mode.)

This fixes the storage performance tests, which run on the
garnet-x64-perf-dawson_canyon builder.

Bug: 23014
Test: garnet-x64-perf-dawson_canyon and peridot-x64-perf-dawson_canyon trybots
Change-Id: Idab8b9c02dca6e938f37528b7e6036c9260636d7