[cmc] Use stronger types for references and names.
Currently we use "String" for references (such as "realm" or "#child"),
for names (such as "child"), and for other entities (such as paths, and
storage types).

Move the CML validation/compilation code to use stronger types,
distinguishing amongst these. In particular, we parse the raw strings
during deserialization, allowing the rest of the code to use the
stronger types.

No changes in behaviour.

Change-Id: I833c3d42907fccb5ebfbf982081d1b073395c669