[kernel][vm] Add markers to VmPageList
Changes VmPageList to be able to store a 'marker' at an offset, instead
of just a vm_page*. For the VmPageList a marker is treated nearly
identically to an actual page and iteration helpers such as
ForEveryPage will trigger on both markers and pages. Similarly when
page lists are merged or spliced, markers are transfered along with the
actual pages.

To simplify the API a wrapper type called VmPageOrMarker is used that
ensures vm_page's are not leaked accidentally, and that pages and
markers do not get confused. Internally the implementation uses a
sentinal value in a vm_page* to indicate a marker and so should have
no space or runtime overhead.

Majority of this change involves plumbing the new wrapper type
through all the usages of page lists and adding any logic for handling
'pages that are not really pages'. This logic is generally trivial in
the context of the operation being attempted.

Test: The VmPageList unit tests are altered, both for the API type
changes, and to test that markers are manipulated properly. The
existing core tests validate that existing behavior of VmObjetPaged has
not changed.

Currently there is no way to cause a VmPageList in a VmObjectPaged to
have markers and so there are no tests for those code paths. These
tests will be added with the zero operation when markers are able to be
produced.

Bug: 36474

Change-Id: Ia2161a457c09b0937ec7e443a9292dd4c5e12243