[fidl][rust] Refactor fidl::encoding for transitional wire format
This CL refactors the fidl::encoding module to avoid duplication of
inline_align and inline_size functions in the Encodable and Decodable
traits. These two functions now live in a separate trait called Layout,
and they take a new parameter to enable context-sensitive
implementations.

The purpose of this refactoring is to ease the implementation of a
transitional wire format for the union-to-xunion migration. During this
migration, we must have a transitional stage where unions are sometimes
encoded/decoded using the xunion format based on build flags or
transaction header flags.

The reason the duplication existed in the first place is that we
sometimes use dynamic dispatch via Encodable trait objects. This helps
reduced binary bloat, but it complicates code-sharing with Decodable,
which is not object-safe (cannot have trait objects). This CL preserves
the ability to use Encodable trait objects, with some restrictions noted
in the Encodable documentation comment.

Test: fx run-test rust_fidl_tests

Bug: 38451
Change-Id: I20d35bdd4d8888131cde8fa84d0b3c0a159137fe