[fuzzing][fs] add fuzzer for vmo_file
This CL adds 'vmo_file-fuzzer' to fuzz the vmo-file library.

The fuzzer uses the fuzzer data to create a vmo of a random size (up to
40 MB), and uses that to create a VmoFile.

After creating the VmoFile, the fuzzer will try to write some data to
it, and if it succeeds, read it back.

Finally, the fuzzer calls GetNodeInfo on the VmoFile.

Change-Id: If4562f6a0fff997447225769d6e835ca1bd0ddd7