[routing] Add routing.fidl
This CL defines APIs for routing table management, and serves
the communication between Network Manager and Netstack*. Network
manager is to be the sole privileged in the end. (single writer model).
Other components such as net-cli are allowed to use this protocol
during the migration phase.

- Route and RouteProperties are defined separately.
- Route is the key of RouteTableEntry.
  Multiple defaults are supported.
  Multiple destination subnets with different egress is supported.
  One may delete a RouteTableEntry by specifying the Route.
- RouteProperties may be updated without deleting and adding back
  a Route.
- ChangeRouteTable() may change multiple RouteTableEntry as a single
  commit.
  It is all-or-nothing model.
- Get/SetInterfaceMetric() supports today's design on route metric
  association.

This CL is not intended to include implementation of this protocol.
Comments are added to netstack.fidl to assist reviewers. Those are not
to be landed in this CL.

Change-Id: I0309099ddf8370b8bbb409379eaf55724cedfed6