[activity] Implement fuchsia.ui.activity.Control
The Control API is a transitional interface which can be used to
override the activity state determined by the Activity Service.

Once a state is assigned through the Control API, listeners will only be
notified of state transitions which happen through this Control API
(unless the Activity Service restarts). The Tracker API will continue to
handle input as always.

This CL also includes a small cleanup to remove locks from
activity_provider_connection.cc (Activity Service is single-threaded so
these locks are unnecessary).

Bug: 38334
Change-Id: I551bb737375075bb597ba99f4ff739bcf8fa94eb