[inspect] Detect cycles in extents and do not preallocate.
We could end up getting an OOM due to preallocating the size of a
property based on the length in the file. Since this can be up to 2^32,
we could use all available memory.

Also, we did not check for cycles in linked extents, meaning we could
end up with an infinite (up to maximum memory) loop.

This change catches cycles, which means that the maximum size of a
property is less than the total size of the Snapshot it is being read
from.

Bug: 39177
Change-Id: I06c613590a50d8f30af2581b1122c0daa5bf9eee