[camera] Removing fuchsia.camera.common (Part 1/2)
fuchsia.camera.common is an illegally named fidl library.
We could rename it, but since it is a duplicate of what is
in the fuchsia.camera library, we are just deleting it.
To avoid breaking existing code, this change is split into
two parts:

Part 1: Rename all fuchsia.camera.common usages to camera
Part 2: Remove fuchsia.camera.common

Test: This change does not introduce any functional changes.

Change-Id: Ib3421df396c056d513d7b532cb6564e8f0a93217