[fx] run -> qemu; aemu -> emu
We want to use `fx run` in the future as an alias for run a component, and we
want to use `fx emu` as the canonical emulator in the future. As such we are
re-arranging the commands.

By November we can increase the timeout to 10s+.
By December the aliases will be removed.

Additionally:
net-run is removed. It has now OWNER listed, but the original author and
other contributor both reported they have not used it in a long time.

A large scale doc change to update all instances of "fx run" with one of "fx
emu" or "fx qemu".

Change-Id: I31654152640c78dcf923a7356dcda69b10a45afd