[thread-safe-deleter] Delete on correct thread.
Fuchsia has some types such as FIDL callbacks that are affinitized to a
thread, but also safe to curry to other threads (and back).  This class
is a helper to ensure that deletion occurs on the affinity thread even
in cases where things are shut down abrubtly and deleted on the wrong
thread.

This will be used by a later CL that curries a FIDL callback using a
ThreadSafeDeleter.

Change-Id: I8358a6c5260b9da54b7cc0458cf03eb4dbcafff3