[fonts][rust] Irons out FFI lifetimes for FT_Open_Args.
The FT_Open_Args does not own its arguments, so ends up
leaking a pathname which had to be passed to it with
an intermediate conversion to a C string.

This changed code wraps the FFI type FT_Open_Args to ensure
that the resulting type always remains live as long as
needed.

Bug: 38228
Change-Id: If81c49643486c459f03962672d5712f067da1bbf