[debugger] Move register querying to the frame.
Previously register querying was on the thread. But each frame in the
unwound stack can provide its saved registers.

This can be more consistent when there is only one way to get the
registers, and it can return either the saved or the current registers
according to what type of frame it is.

This adds support for asynchronously querying the non-general registers.
This shouldn't really be used yet sinc ethe final hook-up for the
expression evaluator is not complete.

TEST=There should generally be no behavior change. The async querying
will be tested in the followup patch when it's actually hooked up.

Change-Id: I364284cc5fc0c9007eb3ce9aa5d4cd195f297706