[media.sessions] Catch new watchers up
A new watcher is not immediately caught up with existing players. Tests
assumed this was the case, but the timing between a watcher channel and
a player publisher's channel with the media session service is not
deterministic.

This resulted in tests hanging sometimes when the service learned about
player before learning about the watchers.

Bug: 39343
TEST: Regression test added; no longer flakes
Change-Id: I4f377477334218333efac01be27135df587d3b18