[network-manager] Basic device schema
This commit adds an initial barebones device model for network
manager. There are many things left to do, but it does add some of
the core interface and subinterface schemas.

This commit has no functional changes: Broke the commit up due to size.

This is based on the OpenConfig YANG models, expressed more simply using
JSON Schema.

The definition of interfaces/subinterfaces are the OpenConfig ones.

Enum values are taken from the OpenConfig data models, which are based
on RFC2863, the "Standard Interfaces MIB".

Bug: 38672
Tests: Ran network_manager unit tests for good measure.

Change-Id: I91cdb9facc7ad559dec12597442e9d269afd2559