[zxcrypt] Wait indefinitely for devices to appear
This change makes the test wait for fvm device infinitely and relies on the test
timeouts if the device does not appear. At the moment FVM bind returns
immediately without the bind actually being completed. Its possible the
flakes are cause because 3 sec time is way too short for the device to
bind and the device that we are waiting on is getting created.

Test:
while /system/test/sys/zxcrypt-test; do true; done
zxcrypt tests in loop successful for 5 min.

Change-Id: I3aaf2abf02e03cd6c079c2a2e4ad6134708b050a