[media][test] Fidelity testing at add'l frequencies
This CL adds four additional test frequencies that specifically
assess the transition-band of our audio resamplers. The metric in
question is the transition-band width -- how narrow is the
transition from pass-band to stop-band, when applying a low-pass
filter during downsampling. Measurements show that the current
LPF implementation has a transition band of approx. 2.5% (for a
nominal sampling rate of 48 kHz, this equates to +/-600 Hz from
the LPF target frequency).

We also change the resampling ratio referred to as 'MicroSRC',
so that it is "ever-so-slightly downsampling" rather than "ever-
so-slightly upsampling", so that we can observe the effects of
the low-pass filter in that usage. For this reason, the set of
expected fidelity results for this rate ratio have changed.

And finally, there is a small amount of code rework intended to
prepare the way for a future CL that will add tests of our
ability to block very low frequencies (DC rejection). Toward
this goal, the const kFirstInBandRefFreqIdx has been introduced,
and the const kNumInBandReferenceFreqs has been changed to
kFirstOutBandRefFreqIdx.

Test: test-only CL, no production code changes
      fx run-test audio_fidelity_tests --full

Change-Id: I85c3fb278199bc5c617dc70aedf84b8c1a719921