Revert "[build][products] Set product_name"
This reverts commit 7a1ad1fbe6fdd7c0988a25a25303ce0f8800b0dc.

Reason for revert: We should not set the product name as a global variable in the build.  I'm sorry that I don't have time to dig up the references for when we've discussed this topic previously, but this CL was not reviewed by the //product/OWNERS and should be reverted.

Original change's description:
> [build][products] Set product_name
> 
> Set product_name for all products, and set
> build_info_product to product_name. We already
> do this for board_name and build_info_board.
> 
> This change will allow us to stop inferring
> build_info_product at recipe runtime.
> 
> Tested locally, and is also integration tested
> by all tryjobs. Does not introduce any functional
> change on bots while fxr/331708 is unsubmitted.
> 
> Bug: 10247
> Change-Id: I89f314fbb54f65dec411e9208384307c18e641af

TBR=raggi@google.com,joshuaseaton@google.com,atyfto@google.com

Change-Id: I380c44051dbc8ae8c242a654d5c37e8903ee0455
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: 10247