[camera] Removing fuchsia.camera.common (Part 2/2)
fuchsia.camera.common is an illegally named fidl library.
We could rename it, but since it is a duplicate of what is
in the fuchsia.camera library, we are just deleting it.
To avoid breaking existing code, this change is split into
two parts:

Part 1: Rename all fuchsia.camera.common usages to camera
Part 2: Remove fuchsia.camera.common

Test: This change does not introduce any functional changes.

Change-Id: If447b3e9beca3e2b7b5423e06557c21f128cff3b