[inspect][validator] Diff mode
Add a "diff mode" with command line switch. When trees differ, you
can ask for full render of the trees, a compressed render of only the
differing lines (which may still be large), or both.

Test: README.md
Also, I modified a test to fail and tested --difftype of (none/default), "full",
"diff", "both", "", and "asdf" to make sure output was as expected.

CF-911

Change-Id: If457b1e4aa89427c34ebe81f52d40b12eb45a1ee