[cmc] Remove dead duplicate "to" detection in validator.
The validator currently has code to detect when an "offer" clause has
the same "to" target twice. This can't be exercised or tested, because
the JSON schema already prevents duplicate values appearing in the "to"
clause.

Add a test for a duplicate "to" target in an "offer" statement, and
remove the dead code.

Change-Id: I874ef4b81743b36f099b2a7960023881215444fb