[magma] Allow creating a device with a channel
The current entrypoints take a file descriptor and use fdio to get a
channel from it. We want to avoid using fdio in the magma ICD, so add a
new entrypoint that creates a magma_device_t from a channel, and use the
magma_device_t in all the other entrypoints.

Test: new magma_abi_conformance_tests
Bug: 13095
MULTIPLY: `[
  {
    "target": "//garnet/lib/magma/tests:magma_abi_conformance_tests",
    "total_runs": 30
  }
]`

Change-Id: I7a2bb1737ec5bd655bbf0f86a048cc6789b349b3