[fs][journal] update vmo offset in journal replay tree
The update function for the buffered operation range container now
updates the vmo offset relative to the dev_offset instead of being
updated to the minimum vmo offset value.

The update function was updating the vmo_offset by taking the minimum of
the two offsets when merging two entries. This is the right thing when
merging two entries, but if a new range is placed in the tree that
updates a part of an existing range, and that range gets broken apart,
_and_ the updated entry is earlier than any other operations in the same
range, then the other operations are updated with vmo offset of the
updated range instead of their own.

We know that for a given set of merged operations, the difference
between their respective offsets is stable - this is guaranteed by check
in the static update function we provide. Essentially, the vmo offset
and the dev offset must both be valid ranges in the set of operations.
We use this property to update the vmo offset when we update the dev
offset - however far apart they are, after an update, they should stay
that far apart.

This behavior caused corruptions in the filesystem in extremely specific
circumstances. This type of merge/break-up behavior only manifested
itself when the first and the second block of the inode table were both
updated, in order, in a single transaction, and then the first block in
the inode table was updated again later, but the second block wasn't
updated again before the power failure event (which would've caused that
block to be corrected), which causes the second block of the inode table
to be overwritten by an old version of the first during replay.

This also adds a unit test case which tests the replay tree in a way
that triggers this behavior. This test case failed before the fix here,
and passes after.

Test: /system/test/fs/journal-unit-test
Bug: 39170
Change-Id: Ibc620fad8fddc36c978280fa4f67c9dae7a0244f